--- frontend/src/metabase/containers/search/components/CommonPage/Index.css
@@ -1,21 +1,0 @@
-.dashboards__web-content {
-  text-overflow: ellipsis;
-  overflow: hidden;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  color: var(--footprint-color-primary-text);
-}
-
-.dashboards__web-url {
-  margin: 6px 0;
-  color: var(--footprint-color-secondary-text1);
-}
-.search .dashboards__list-item:hover .dashboards__web-title,
-.search .common-page__title:hover {
-  text-decoration: underline;
-  text-decoration-color: #3434b2;
-}
-
-.search .dashboards__table {
-  margin: 0;
-}
